A lot of stuff:
* Port to .NET 6
** Remove support for Soap serialization
** Move stuff from AssemblyInfo.cs over to csproj file
** Remove Compile and EmbeddedResource tags from csproj file
** Remove unnecessary references from csproj file
** Removed .NET Framework requirement from help text (We can ship self-contained executables that include the necessary runtime), and also the disk space requirement temporarily until I calculate how much is necessary for the self-contained executable

Note: In the process I also removed some stuff from the csproj file that seemed to me redundant or unnecessary (e.g. ProjectGuid, FileAlignment...). If I was mistaken I'll put it back in a future commit :).

* Remove DirectX support (as well as all DirectX related files, code, and references to DirectX in help files or code comments)
* Remove solution file and all references to RSConfig (all options should be settable in game, why complicate things?). I added the options to change window mode and toggle writing to a log file into the in-game options menu. I didn't add options to switch video/audio mode, since the goal is to have one cross-platform handler for both, once the GDI -> SFML port is done.
* Stop distributing SFML, use the NuGet package (also bumped the version to the latest: 2.5.1)

I did all this stuff in one commit because a lot of it was necessary to get this to even build with my development environment (.NET SDK 6.0.200 on Windows 7)... but then I got carried away and kept doing other stuff and was too lazy to split it into clean commits.